+++ api/server/information/.gitignore
