+++ api/server/information/.golangci.yaml
