+++ api/server/information/internal/application/auth.go
@@ -3,7 +3,7 @@ package application
 import (
 	"context"
 
-	"github.com/calmato/gran-book/api/server/notification/internal/domain/auth"
+	"github.com/calmato/gran-book/api/server/information/internal/domain/auth"
 )
 
 // AuthApplication - Authアプリケーションのインターフェース
