+++ api/server/information/internal/domain/auth/service.go
