+++ api/server/information/internal/domain/exception/error.go
