+++ api/server/information/internal/domain/exception/validation.go
