+++ api/server/information/internal/infrastructure/service/auth_test.go
@@ -1,15 +1,16 @@
-package application
+package service
 
 import (
 	"context"
 	"reflect"
 	"testing"
 
-	mock_auth "github.com/calmato/gran-book/api/server/notification/mock/domain/auth"
+	"github.com/calmato/gran-book/api/server/information/internal/domain/exception"
+	mock_auth "github.com/calmato/gran-book/api/server/information/mock/domain/auth"
 	"github.com/golang/mock/gomock"
 )
 
-func TestAuthApplication_Authentication(t *testing.T) {
+func TestAuthService_Authentication(t *testing.T) {
 	testCases := map[string]struct {
 		Expected struct {
 			UID   string
@@ -25,6 +26,15 @@ func TestAuthApplication_Authentication(t *testing.T) {
 				Error: nil,
 			},
 		},
+		"ng_unauthorized": {
+			Expected: struct {
+				UID   string
+				Error error
+			}{
+				UID:   "",
+				Error: exception.Unauthorized.New(nil),
+			},
+		},
 	}
 
 	for result, tc := range testCases {
@@ -34,11 +44,11 @@ func TestAuthApplication_Authentication(t *testing.T) {
 		ctrl := gomock.NewController(t)
 		defer ctrl.Finish()
 
-		asm := mock_auth.NewMockService(ctrl)
-		asm.EXPECT().Authentication(ctx).Return(tc.Expected.UID, tc.Expected.Error)
+		arm := mock_auth.NewMockRepository(ctrl)
+		arm.EXPECT().Authentication(ctx).Return(tc.Expected.UID, tc.Expected.Error)
 
 		t.Run(result, func(t *testing.T) {
-			target := NewAuthApplication(asm)
+			target := NewAuthService(arm)
 
 			got, err := target.Authentication(ctx)
 			if !reflect.DeepEqual(err, tc.Expected.Error) {
