+++ api/server/information/lib/array/convert.go
