+++ api/server/information/lib/firebase/client.go
